refactor(orders): decompose Manager OrdersController and add ServiceRegistry factories
Extract list/mutation/products/presenter services to shrink OrdersController
below 1k LOC; register manager order services in DI. Replace ServiceRegistry
switch/in_array wiring with explicit ServiceRegistryFactories map.
